--- AEP-Assurance/Assets/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0949bd4d57d484141b78c6d6812fb497
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
