--- AEP-Assurance/Assets/Editor/BuildPostProcessor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d9263422159164dccb0941f7bef652fa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
