+++ AEP-Assurance/Assets/Plugins/iOS/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0949bd4d57d484141b78c6d6812fb497
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
